+++ packages/ui/src/main.ts
@@ -18,9 +18,9 @@ import HerodotusFormEngine from '@herodotus-cloud/form-engine';
 import DisableDevtool from 'disable-devtool';
 
 import { setupStore, setupQuasar, setupI18n, setupUploader } from '@/plugins';
-import router, { setupRouter } from './routers';
+import { setupRouter } from './routers';
 
-import { CONSTANTS, VARIABLES, setupKernel } from '@/configurations';
+import { CONSTANTS, VARIABLES } from '@/configurations';
 
 if (typeof (window as any).global === 'undefined') {
   (window as any).global = window;
@@ -42,8 +42,6 @@ async function setupApp() {
 
   // 挂载路由
   await setupRouter(app);
-
-  setupKernel(router);
 
   setupUploader(app);