+++ shared/framework-kernel/src/lib/hooks/index.ts
@@ -1,0 +1,3 @@
+import useEditFinish from './useEditFinish';
+
+export { useEditFinish };