+++ shared/framework-kernel/src/lib/hooks/useEditFinish.ts
@@ -1,18 +1,18 @@
 import { useRoute } from 'vue-router';
-import { useTabsStore, useRouterStore } from '@/stores';
-import { RouterUtils } from '@/configurations';
+import { useTabsViewStore, useRouterStore } from '../stores';
+import { RouterUtilities } from '../utilities';
 
 export default function useEditFinish() {
   const route = useRoute();
   const routeStore = useRouterStore();
-  const tabs = useTabsStore();
+  const tabs = useTabsViewStore();
 
   const onFinish = () => {
     const name = route.name as string;
 
     routeStore.removeRoutePushParam(name);
     tabs.deleteTab(route);
-    RouterUtils.goBack();
+    RouterUtilities.getInstance().goBack();
   };
 
   return {