+++ shared/framework-kernel/src/lib/index.ts
@@ -1,3 +1,4 @@
-export * from './utilities';
+export * from './hooks';
 export * from './stores';
+export * from './utilities';
 export * from './main';