+++ shared/components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @herodotus-cloud/components
 
+## 1.6.5
+
+### Patch Changes
+
+- [重构] 迁移前端 useEditFinish 代码至模块 @herodotus-cloud/framework-kernel
+- Updated dependencies
+  - @herodotus-cloud/core@1.6.5
+
 ## 1.6.4
 
 ### Patch Changes